requirements.txt
streamlit==1.34.0
tensorflow==2.17.0
Pillow==9.4.0
numpy==1.26.4
streamlit-drawable-canvas==0.9.3
